Made dry update an option on the normal update command.
