+++ GitDepend.UnitTests/GitDepend.UnitTests.csproj
@@ -85,7 +85,6 @@
     <Compile Include="Commands\CleanCommandTests.cs" />
     <Compile Include="Commands\CloneCommandTests.cs" />
     <Compile Include="Commands\CommandParserTests.cs" />
-    <Compile Include="Commands\DryUpdateCommandTests.cs" />
     <Compile Include="Commands\InitCommandTests.cs" />
     <Compile Include="Commands\ListCommandTests.cs" />
     <Compile Include="Commands\ManageCommandTests.cs" />
