+++ GitDepend/CommandLine/Options.cs
@@ -39,9 +39,6 @@ namespace GitDepend.CommandLine
         [VerbOption(ConfigCommand.Name, HelpText = "Displays the full configuration file")]
         public ConfigSubOptions ConfigVerb { get; set; } = new ConfigSubOptions();
 
-		[VerbOption(DryUpdateCommand.Name, HelpText = "Executes a dry run of the update command")]
-		public UpdateSubOptions DryUpdateVerb { get; set; } = new UpdateSubOptions();
-
 		[VerbOption(InitCommand.Name, HelpText = "Assists you in creating a GitDepend.json")]
         public InitSubOptions InitVerb { get; set; } = new InitSubOptions();
