+++ GitDepend/CommandLine/UpdateSubOptions.cs
@@ -14,5 +14,11 @@ namespace GitDepend.CommandLine
         /// </summary>
         [Option('f', "force", DefaultValue = false, HelpText = "Forces dependencies to build before updating packages.")]
         public bool Force { get; set; }
-    }
+
+		/// <summary>
+		/// Executes a dry run update. Processes needed updates without actually executing them.
+		/// </summary>
+		[Option('d', "dry", DefaultValue = false, HelpText = "Checks for required updates without executing them.")]
+		public bool Dry { get; set; }
+	}
 }
