+++ GitDepend/GitDepend.csproj
@@ -129,7 +129,6 @@
     <Compile Include="Commands\CleanCommand.cs" />
     <Compile Include="Commands\CloneCommand.cs" />
     <Compile Include="Commands\CommandParser.cs" />
-    <Compile Include="Commands\DryUpdateCommand.cs" />
     <Compile Include="Commands\ICommand.cs" />
     <Compile Include="Commands\InitCommand.cs" />
     <Compile Include="Commands\ConfigCommand.cs" />
